requirements.txt
python-slugify==8.0.1
pytube==15.0.0
text-unidecode==1.3
typing_extensions==4.9.0
